% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fold.R
\name{folding.ratio}
\alias{folding.ratio}
\title{Computes the folding ratio of the input data}
\usage{
folding.ratio(X)
}
\arguments{
\item{X}{nxd matrix (n observations, d dimensions)}
}
\value{
the folding ratio
}
\description{
Computes the folding ratio of the input data
}
\examples{
X = matrix(runif(n = 1000, min = 0., max = 1.), ncol = 1)
phi = folding.statistics(X)

}
