% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fold.R
\name{folding.statistics}
\alias{folding.statistics}
\title{Computes the folding statistics of the input data}
\usage{
folding.statistics(X)
}
\arguments{
\item{X}{nxd matrix (n observations, d dimensions)}
}
\value{
the folding statsistics
}
\description{
Computes the folding statistics of the input data
}
\examples{
library(MASS)
mu = c(0,0)
Sigma = matrix(c(1,0.5,1,0.5), ncol = 2)
X = mvrnorm(n = 5000, mu = mu, Sigma = Sigma)
Phi = folding.statistics(X)

}
