% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fold.R
\name{folding.test.bound}
\alias{folding.test.bound}
\title{Computes the confidence bound for the significance level p}
\usage{
folding.test.bound(n, d, p)
}
\arguments{
\item{n}{sample size}

\item{d}{dimension}

\item{p}{significance level (between 0 and 1, the lower, the more significant)}
}
\value{
the confidence bound q (the bounds are 1-q and 1+q)
}
\description{
Computes the confidence bound for the significance level p
}
\examples{
n = 2000  	# number of observations
d = 2 	  	# 2 dimensional data
p = 0.05  	# we want the bound at the level 0.05 (classical p-value)
q = folding.test.bound(n,d,p)

}
